refactor: remove naive utils file
